Missing Data and Imputation
Missing data are a common problem in many fields of human endeavor
ranging from social
sciences to economics and from political research to entertainment
industry.  The objective of
this  thesis  is  to  study  and  apply    missing  data  algortihms
for  treatment  of  missing  data
which encountered on bank marketing data sets.  Six missing data
imputation algortihms are
analyzed  which  are  mean,  hot-deck,  k-nearest  neighbor,
regression,  expectation
maximization and multiple imputation method. By imputation of missing
data, the accuracy
of clustering and association with regard to aspect of the data can be
increased.
This thesis also explores the advantages and disadvantages of each used
imputation method.
Similarities and differences between each imputation method is
discussed, as well